src/main/java/de/ugoe/cs/seppelshark/configuration/Database.java
/*
 * Copyright (C) 2017 University of Goettingen, Germany
 * 
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *         http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package de.ugoe.cs.seppelshark.configuration;

import com.google.common.base.MoreObjects;

/**
 * @author Fabian Trautsch
 */
public class Database {
    private String hostname = "localhost";
    private int port = 27017;
    private String database = "smartshark";
    private String username = null;
    private String password = null;
    private String authenticationDatabase = null;

    public Database() {}

    public String getHostname() {
        return hostname;
    }

    public void setHostname(String hostname) {
        this.hostname = hostname;
    }

    public int getPort() {
        return port;
    }

    public void setPort(int port) {
        this.port = port;
    }

    public String getDatabase() {
        return database;
    }

    public void setDatabase(String database) {
        this.database = database;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getAuthenticationDatabase() {
        return authenticationDatabase;
    }

    public void setAuthenticationDatabase(String authenticationDatabase) {
        this.authenticationDatabase = authenticationDatabase;
    }

    public String toString() {
        return MoreObjects.toStringHelper(this)
                .add("hostname", hostname)
                .add("port", port)
                .add("database", database)
                .add("username", username)
                .add("password", password)
                .add("authenticationDatabase", authenticationDatabase)
                .toString();
    }



}
